v1.4
Version 1.4 (28925 conventional, 5296344 experimental)

Added ghost snare hits on even quarter-beats with minimal constraints.

The solve notebook has been updated to display and write the new hits. Ghost snare hits are the same as snare hits with a reduced velocity of 60 instead of 100.

MIDI pattern examples are included in this version.