v1.4.1
Version 1.4.1 (2707 conventional, 506064 experimental)

Placement of the ghost snares are much more constrained - especially around other snares. This resulted in a pleasing improvement to the groove of each pattern.

The velocity of the ghost snares has been reduced to 40 in the solve notebook.

More example patterns are included in this version.