v1.5
Version 1.5 (Same number of patterns)

Significant code restructuring in this version.

The necessary time intervals are now written as facts instead of being generated with rules. They reside in the time.lp ruleset.

There are seperate rulesets for the placement of each drum hit and for additional constraints on top of those placements. Level 1 constraints translate to being experimental and conventional constraints. Level 2 constraints are only for conventional patterns. If there is a choice between conventional and experimental placement of a hit, only one ruleset should be picked.