v1.7
Version 1.7 (Very large amount of possible patterns)

Functionality has been added to allow two bar drum patterns to be generated. This is achieved through feeding the random solution from the one bar problem into another problem to be solved. This problem has a seperate rule set, alongside reusing some of the one bar rule sets, to determine if any modifications to the original pattern should take place for the two bar version.

Changes have been made to the solve notebook to accomodate this. In particular, the print_hits method scales with the length of the pattern. Also a new method called extend_pattern has been added to extend the one bar pattern to a given length. However it only performs extension to two bar patterns for now. The generate_patterns method now includes a while loop to continue to search for valid two bar loops if it doesn't manage to find one first time. It does this by changing the random one bar pattern to a different solution before extending it, if no two bar solutions could be found to the previous one.

Example MIDI patterns can be found in the v1.7 Patterns folder under the new Example Patterns folder.

The rules for two bar extension will be refined in future versions.