v1.8
Version 1.8 (practically infinite different patterns)

Mild random humanisation of the drum patterns has been developed. This is in the form of a humanisation_intensity parameter that controls how much hits are randomly nudged. A normal distribution with mean 0 models the nudge. The humanisation_itensity value forms the standard deviation of the distribution. The velcoity of the nudged notes are varied relative to the nudge amount too (The greater the nudge the quieter the hit).

In the conventional snare placement rule set, the option to have a snare on the 3rd quarter-beat of the measure has been removed.