v1.8.2
Version 1.8.2

After beginning work on the UI using PyQt5, the package used has since changed to PyQt4 for ease of use.

The solve notebook has been converted to composer.py module using the notebook_to_py script. This is imported into auto_drum_comp.py, which contains the start of the UI. This script also contains the main method to run the application.

qt_designer.ui is the Qt designer file which has generated a lot of the UI code as a time saver.

So far only the check boxes impact the patterns generated by the app. They include each hit checked in the output pattern.

Implementing the sliders and neatening the design will be next on the agenda.