v1.8.4
Version 1.8.4

All configurable parameters can be inputted by the user through the user interface. A dial is used for humanisation amount; a drop down menu (combobox) is used for the pattern length; a scrollbox is used for the amount of patterns generated and a textbox is used to input additional drum hits.

This general layout can still be improved specifically the way user input additional hits.

The problem with the humanisation logic still needs to be corrected.