v1.8.5
Version 1.8.5

A file menu has been implemented. This contains a basic quit options and also an option to choose the save path of the generate MIDI file(s).

The layout of the app has been split into three tabs named: basic, advanced and pattern input. The basic tab only contains pattern length and number of pattern parameters as well as a text box to enter the file save name and the generate button. All other parameters apart from the user input text box are found under advanced tab.

When the generate button is clicked the first pattern generated appears underneath the button. The colourmapping representing hit velocity has been improved (the lighter the colour the lower the MIDI velocity value).

The user input method could still be improved (with it looking more like an interactive gird similar to what is displayed when a pattern is generated). Also it is worth investigating whether audio of the pattern could be played after generation.