v1.9
Version 1.9

The drum patterns generated can now be played as audio with the play button. This uses fluidsynth to virtually play the MIDI with the dnb_kit Soundfont. Midi2audio contains methods to make this process simpler.

The number of solutions generated is now limited by the time taken to solve the model. It has been set to 10s for the meantime.

Improving the backend of the app is the next task on the list. Extending the pattern length arbitrarily as well as including a drum fill method will be the first additions to make.