v2.1
Version 2.1

Simple fills have been added at the end of even numbered bars. These occur randomly with either a short fill or no fill being played at the end of a multiple of 2 or a long, short or no fill being played at the end of a multiple of 4.

These fills are random at the moment - using simple placement in the short_fills.lp and long_fills.lp files. These will be improved in future versions.