v2.2
Version 2.2

Fills have been improved and are slightly less random. There is an option to toggle them on and off in the UI.

Fill hits and ordinary hits have been classified to extend patterns better. This allows us to stop the more random fill sections influencing the next bar generated.

A initial distribution build has been generated using PyInstaller. It relies on having the rule sets and clingo solver bundled in with it for now. Also, there is no way to playback audio as of yet.