v2.2.4
Version 2.2.4

Some of the rule sets have been refactored slightly for readability.

A new distribution has been made. The play button is still unlikely to work on other computers and the build does not work as a single file on its own. The reason for this is still unknown. The rules, solver, soundfont and app icon have to be bundled with the distribution presently.