v2.4
Version 2.4

Source code is now multi-platform. This mainly involved creating a platform specific separator variable to ensure compatibility on all platforms.

Clingo is imported as a package now. There is no need for the clingo.exe solver. This also improves cross compatibility.

The FluidSynth executable has also been removed as it is not compatible with all versions of Windows. Users will have to install FluidSynth themselves if they want to use the 'Play' feature.