Track users hit preferences
By creating a Tracker object in the composer we can keep track of each hit in a pattern the user likes or dislikes.

Like and Dislike button will appear once a pattern is generated. Clicking these buttons will influence the rating stored in a database for each hit contained in the pattern. The key stored in the database is the [hit, time] pairing outputed by the composer. This will make it easier to retrieve them later.