Composer learns user's preferences over time
A simple epsilon decreasing algorithm has been implemented. It will continuously learn down to a minimum learning rate of epsilon=0.5 over 500 trials.

Hit, time data collected is condensed to a one bar pattern for preference storage.

The ratings stored are converted into probability weighting and used to sample 20 hits to form a new one bar pattern when desired. This needs to be extended to the number of bars desired. This will be done initially by sampling n bars separately and multiplying out the time appropriately.

The number of hits to sample also needs to be intelligently determined. Fixing to 20 is not sensible necessarily.